Split scratchstack-database ops/iam/policy up
