+++ scratchstack-database/src/ops/iam/mod.rs
@@ -6,7 +6,6 @@
 
 use {
     crate::constants::iam::*,
-    scratchstack_arn::Arn,
     scratchstack_pagination::{
         FixedKeyService, OperationPaginator, ScratchstackOperationMetadata, ScratchstackServiceMetadata,
     },
@@ -14,7 +13,6 @@ use {
         error_meta::Error as IamError,
         types::error::{InternalFailure, ValidationError},
     },
-    std::str::FromStr as _,
 };
 
 mod account;
@@ -208,68 +206,4 @@ pub fn constrain_max_items(max_items: Option<i32>) -> Result<usize, ValidationEr
     } else {
         Ok(100)
     }
-}
-
-/// Parts extracted from a policy ARN.
-pub(crate) struct PolicyArnParts {
-    /// Owning account id for the policy. The literal `aws` alias is normalized to
-    /// `000000000000`.
-    pub(crate) account_id: String,
-    /// Canonical policy path, including a leading slash.
-    pub(crate) policy_path: String,
-    /// Lowercase policy name for case-insensitive lookup.
-    pub(crate) policy_name_lower: String,
-}
-
-/// Parse a policy ARN and extract the account id, path, and lowercase policy name. Returns a
-/// `ValidationError` if the ARN is unparseable or does not point at a `policy/` resource.
-pub(crate) fn parse_policy_arn(policy_arn: &str) -> Result<PolicyArnParts, IamError> {
-    let arn = Arn::from_str(policy_arn).map_err(|e| {
-        log::info!("Failed to parse policy ARN: {e}");
-        IamError::from(ValidationError::builder().message("Invalid policy ARN".to_string()).build())
-    })?;
-
-    let service = arn.service();
-    if service != ARN_SERVICE_IAM {
-        return Err(ValidationError::builder()
-            .message("Policy ARN must have service \"iam\"".to_string())
-            .build()
-            .into());
-    }
-
-    if !arn.region().is_empty() {
-        return Err(ValidationError::builder().message("Policy ARN must not have a region".to_string()).build().into());
-    }
-
-    let resource = arn.resource();
-    if !resource.starts_with(ARN_RESOURCE_PREFIX_POLICY) {
-        return Err(ValidationError::builder()
-            .message("Policy ARN must have a resource that starts with \"policy/\"".to_string())
-            .build()
-            .into());
-    }
-
-    let account_id = match arn.account_id() {
-        AWS_ACCOUNT_ID => AWS_ACCOUNT_ID_NUMERIC.to_string(),
-        other => other.to_string(),
-    };
-
-    let policy_path_and_name = &resource[ARN_RESOURCE_PREFIX_POLICY.len()..];
-    let name_start = policy_path_and_name.rfind('/').map(|i| i + 1).unwrap_or(0);
-    let policy_name_lower = policy_path_and_name[name_start..].to_ascii_lowercase();
-    let policy_path = if name_start == 0 {
-        "/".to_string()
-    } else {
-        format!("/{}", &policy_path_and_name[..name_start])
-    };
-
-    if policy_name_lower.is_empty() {
-        return Err(ValidationError::builder().message("Policy name must not be empty".to_string()).build().into());
-    }
-
-    Ok(PolicyArnParts {
-        account_id,
-        policy_path,
-        policy_name_lower,
-    })
 }
